リファクタリング (課題ファイルを別モジュールに分離)
