+++ src/Education/MakeMistakesToLearnHaskell/Exercise/Ex02_5.hs
@@ -1,0 +1,11 @@
+module Education.MakeMistakesToLearnHaskell.Exercise.Ex02_5
+  ( exercise2_5
+  ) where
+
+#include <imports/external.hs>
+
+import Education.MakeMistakesToLearnHaskell.Exercise.Core
+import Education.MakeMistakesToLearnHaskell.Exercise.Types
+
+exercise2_5 :: Exercise
+exercise2_5 = Exercise "2.5" noVeirificationExercise
